[~] Refactoring EmployeeInfoRemoteDataSource.kt - Change logic for showing employee. Now it can work with some null fields. - Moved the model files to a common folder
