+++ tv-app/effecticeTV/app/src/main/java/band/effective/office/tv/di/EmployeeInfoModule.kt
@@ -1,6 +1,6 @@
 package band.effective.office.tv.di
 
-import band.effective.office.tv.domain.models.Employee.EmployeeInfoRepository
+import band.effective.office.tv.domain.model.notion.EmployeeInfoRepository
 import band.effective.office.tv.network.notion.EmployeeInfoRepositoryImpl
 import dagger.Binds
 import dagger.Module
