+++ tv-app/effecticeTV/app/src/main/java/band/effective/office/tv/screen/components/NoStoriesScreen.kt
@@ -14,9 +14,7 @@ import band.effective.office.tv.R
 import band.effective.office.tv.screen.navigation.Screen
 
 @Composable
-fun NoStoriesScreen(
-
-) {
+fun NoStoriesScreen() {
     Surface(
         modifier = Modifier.fillMaxSize(),
         color = MaterialTheme.colors.background,
@@ -30,19 +28,6 @@ fun NoStoriesScreen(
                 fontSize = 26.sp,
                 color = Color.White
             )
-            Button(
-                modifier = Modifier.padding(top = 24.dp),
-                colors = ButtonDefaults.buttonColors(
-                    backgroundColor = MaterialTheme.colors.primaryVariant,
-                    contentColor = Color.White
-                ),
-                onClick = {/* TODO: (Parkhomenko Egor): Need to add navigation */}) {
-                Text(
-                    fontSize = 16.sp,
-                    text = stringResource(id = R.string.back),
-                    modifier = Modifier.padding(horizontal = 24.dp)
-                )
-            }
         }
     }
 }
